requirements.txt
torch>=1.10.0
torchvision>=0.11.0
timm>=0.6.0
numpy>=1.20.0
Pillow>=8.0.0
requests>=2.25.0
